requirements.txt
pyyaml
transformers
sentence-transformers
torch
